# openapi.yaml
...
openapi: 3.1.0
info:
  title: PersonaFlow
  version: 0.1.0
paths:
...
  "/api/v1/rag/ingest":
    post:
      tags:
        - RAG
      summary: Ingest files to be indexed and queried.
      description: |-
        Upload files for ingesting using the advanced RAG system with unstructured library. <br>
                      This API is agnostic of the assistant and includes a /query api to query the indexes. <br>
                      Files can either be provided as an array of URLs within the payload or uploaded via multipart form data. <br>
                    *Note: the payload string should be an IngestRequestPayload (see app.schema.rag.IngestRequestPayload)*
      operationId: ingest_data_from_files
      requestBody:
        content:
          multipart/form-data:
            schema:
              "$ref": "#/components/schemas/Body_ingest_data_from_files"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Ingest Data From Files
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
        - APIKeyHeader: []
  "/api/v1/rag/query":
    post:
      tags:
        - RAG
      summary: Query documents
      description:
        Query ingested documents using advanced RAG system with unstructured
        library. <br>
      operationId: query_documents
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/QueryRequestPayload"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/QueryResponsePayload"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
        - APIKeyHeader: []
...
  "/api/v1/files/{file_id}":
    get:
      tags:
        - Files
      summary: Retrieve file information
      description: Retrieves information about a specific file by its ID.
      operationId: retrieve_file
      security:
        - APIKeyHeader: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: File Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FileSchema"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - Files
      summary: Delete a specific file
      description:
        Deletes a specific file by its ID from the database and file system.
        When a file is deleted, it is also removed from any assistants that may be
        using it and all associated embeddings are deleted from the vector store.
      operationId: delete_file
      security:
        - APIKeyHeader: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: File Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DeleteFileResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    BaseDocumentChunk:
      properties:
        id:
          type: string
          title: Id
        document_id:
          type: string
          title: Document Id
        content:
          type: string
          title: Content
        doc_url:
          anyOf:
            - type: string
            - type: "null"
          title: Doc Url
        source:
          anyOf:
            - type: string
            - type: "null"
          title: Source
        source_type:
          anyOf:
            - type: string
            - type: "null"
          title: Source Type
        chunk_index:
          anyOf:
            - type: integer
            - type: "null"
          title: Chunk Index
        title:
          anyOf:
            - type: string
            - type: "null"
          title: Title
        token_count:
          anyOf:
            - type: integer
            - type: "null"
          title: Token Count
        page_number:
          anyOf:
            - type: integer
            - type: "null"
          title: Page Number
        metadata:
          anyOf:
            - type: object
            - type: "null"
          title: Metadata
        dense_embedding:
          anyOf:
            - items:
                type: number
              type: array
            - type: "null"
          title: Dense Embedding
      type: object
      required:
        - id
        - document_id
        - content
      title: BaseDocumentChunk
    Body_ingest_data_from_files:
      properties:
        files:
          items:
            type: string
            format: binary
          type: array
          title: Files
          description: List of files to upload.
          default: []
        payload:
          type: string
          title: Payload
          description: Ingest request payload as JSON string. (see README for example)
      type: object
      required:
        - payload
      title: Body_ingest_data_from_files
...
    DeleteFileResponse:
      properties:
        summarize:
          type: boolean
          title: Summarize
          description: Create a separate index of document summaries
          default: false
        encoder:
          allOf:
            - "$ref": "#/components/schemas/EncoderConfig"
          default:
            provider: cohere
            encoder_model: embed-multilingual-light-v3.0
            dimensions: 384
        unstructured:
          allOf:
            - "$ref": "#/components/schemas/UnstructuredConfig"
          default:
            partition_strategy: auto
            hi_res_model_name: detectron2_onnx
            process_tables: false
        splitter:
          allOf:
            - "$ref": "#/components/schemas/SplitterConfig"
          default:
            name: semantic
            min_tokens: 30
            max_tokens: 400
            rolling_window_size: 1
            prefix_titles: true
            split_summary: true
      type: object
      required:
        - file_id
        - num_of_deleted_chunks
        - num_of_assistants
        - assistants
      title: DeleteFileResponse
    EncoderConfig:
      properties:
        provider:
          allOf:
            - "$ref": "#/components/schemas/EncoderProvider"
          description: Embedding provider
          default: cohere
        encoder_model:
          type: string
          title: Model Name
          description: Model name for the encoder
          default: text-embedding-3-small
        dimensions:
          type: integer
          title: Dimensions
          description: Dimension of the encoder output
          default: 1536
      type: object
      title: EncoderConfig
    EncoderProvider:
      type: string
      enum:
        - cohere
        - openai
        - huggingface
        - azure_openai
        - mistral
      title: EncoderProvider
    FileSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description:
            A unique identifier for the file. It's a UUID type and is automatically
            generated by the database.
        user_id:
          type: string
          title: User Id
          description: The ID of the user who created the file.
        purpose:
          type: string
          title: Purpose
          description: "The purpose of the file: 'assistants', 'threads', or 'personas'."
        filename:
          type: string
          title: Filename
          description: The preferred name for the file.
        bytes:
          type: integer
          title: Bytes
          description: The file size in bytes, calculated when the file is uploaded.
        mime_type:
          type: string
          title: Mime Type
          description: The mime type of the file.
        kwargs:
          type: object
          title: Kwargs
          description: Any additional kwargs for the file.
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Created date
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: Last updated date
      type: object
      required:
        - id
        - user_id
        - purpose
        - filename
        - bytes
        - mime_type
        - kwargs
        - created_at
        - updated_at
      title: FileSchema
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueryRequestPayload:
      properties:
        input:
          type: string
          title: Input
        index_name:
          type: string
          title: Index Name
        vector_database:
          anyOf:
            - "$ref": "#/components/schemas/VectorDatabase"
            - type: "null"
        encoder:
          allOf:
            - "$ref": "#/components/schemas/EncoderConfig"
          default:
            provider: cohere
            encoder_model: embed-multilingual-light-v3.0
            dimensions: 384
        thread_id:
          anyOf:
            - type: string
            - type: "null"
          title: Thread Id
        enable_rerank:
          anyOf:
            - type: boolean
            - type: "null"
          title: Enable Rerank
          default: false
        interpreter_mode:
          anyOf:
            - type: boolean
            - type: "null"
          title: Interpreter Mode
          default: false
        exclude_fields:
          items:
            type: string
          type: array
          title: Exclude Fields
      type: object
      required:
        - input
        - index_name
      title: QueryRequestPayload
    QueryResponsePayload:
      properties:
        success:
          type: boolean
          title: Success
        data:
          items:
            "$ref": "#/components/schemas/BaseDocumentChunk"
          type: array
          title: Data
      type: object
      required:
        - success
        - data
      title: QueryResponsePayload
...
    SplitterConfig:
      properties:
        name:
          type: string
          enum:
            - semantic
            - by_title
          title: Name
          description: Splitter name, `semantic` or `by_title`
          default: semantic
        min_tokens:
          type: integer
          title: Min Tokens
          description: Only for `semantic` method
          default: 30
        max_tokens:
          type: integer
          title: Max Tokens
          description: Only for `semantic` and `recursive` methods
          default: 400
        rolling_window_size:
          type: integer
          title: Rolling Window Size
          description:
            Only for `semantic` method, cumulative window size for comparing
            similarity between elements
          default: 1
        prefix_titles:
          type: boolean
          title: Prefix Title
          description: Add to split titles, headers, only `semantic` method
          default: true
        split_summary:
          type: boolean
          title: Prefix Summary
          description: Add to split sub-document summary
          default: true
      type: object
      title: SplitterConfig
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    VectorDatabase:
      properties:
        type:
          "$ref": "#/components/schemas/VectorDatabaseType"
        config:
          type: object
          title: Config
      type: object
      required:
        - type
        - config
      title: VectorDatabase
    VectorDatabaseType:
      const: qdrant
      title: VectorDatabaseType
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-KEY